+++ src/base/refl/auto_name.h
@@ -3,11 +3,43 @@
 
 #include <algorithm>
 #include <array>
+#include <functional>
 #include <ranges>
 #include <string_view>
 
-namespace Refl::Name
+namespace Refl
 {
+
+inline namespace Declval
+{
+#ifdef __clang__
+#pragma clang diagnostic push
+#pragma clang diagnostic ignored "-Wundefined-var-template"
+#endif
+
+template <class T> extern T not_exists; // NOLINT
+
+template <class T> consteval T declval()
+{
+	return std::forward<T>(not_exists<std::remove_reference_t<T>>);
+}
+
+#ifdef __clang__
+#pragma clang diagnostic pop
+#endif
+}
+
+namespace Name
+{
+template <class T> struct Wrapper
+{
+	T t;
+};
+
+template <class T> Wrapper(T) -> Wrapper<T>;
+
+template <class Base, class Member>
+static Base getBase(Member Base::*);
 
 template <class E, auto v> consteval auto name()
 {
@@ -15,6 +47,14 @@ template <class E, auto v> consteval auto name()
 	constexpr std::string_view sv = __FUNCSIG__;
 	constexpr auto last = sv.find_last_not_of(" }>)(", sv.size() - 6);
 #else
+#if defined(__clang__)
+	if constexpr (std::is_member_object_pointer_v<decltype(v)>) {
+		return name<void,
+		    Wrapper{&std::invoke(v,
+		        declval<decltype(getBase(v)) &>())}>();
+	}
+#endif
+
 	constexpr std::string_view sv = __PRETTY_FUNCTION__;
 	constexpr auto last = sv.find_last_not_of(" }])");
 #endif
@@ -35,6 +75,12 @@ template <class E, auto v> consteval auto name()
 	else {
 		return std::array<char, 0>{};
 	}
+}
+
+template <auto A>
+constexpr static inline std::string_view in_data_name{std::data(A),
+    std::size(A)};
+
 }
 }