+++ src/chart/main/style.cpp
@@ -437,6 +437,56 @@ Chart Chart::def()
 					.numberScale = ::Text::NumberScale{}
 				},
 			},
+			.subtitle = {
+				Padding
+				{
+					.paddingTop = Gfx::Length::Emphemeral(7.0 / 26.0),
+					.paddingRight = Gfx::Length::Emphemeral(10.0 / 26.0),
+					.paddingBottom = Gfx::Length::Emphemeral(0),
+					.paddingLeft = Gfx::Length::Emphemeral(10.0 / 26.0)
+				},
+				Font
+				{
+					.fontFamily = ::Anim::String(),
+					.fontStyle = Gfx::Font::Style::normal,
+					.fontWeight = Gfx::Font::Weight::Normal(),
+					.fontSize = Gfx::Length::Emphemeral(26.0 / 21.0)
+				},
+				Text
+				{
+					.color = Gfx::Color::RGB(0x494949),
+					.textAlign = Anim::Interpolated<Text::TextAlign>(Text::TextAlign::center),
+					.backgroundColor = Gfx::Color(),
+					.numberFormat = ::Text::NumberFormat::prefixed,
+					.maxFractionDigits = 3,
+					.numberScale = ::Text::NumberScale{}
+				},
+			},
+			.caption = {
+				Padding
+				{
+					.paddingTop = Gfx::Length::Emphemeral(0),
+					.paddingRight = Gfx::Length::Emphemeral(0),
+					.paddingBottom = Gfx::Length::Emphemeral(-1.0),
+					.paddingLeft = Gfx::Length::Emphemeral(5.0 / 12.0)
+				},
+				Font
+				{
+					.fontFamily = ::Anim::String(),
+					.fontStyle = Gfx::Font::Style::normal,
+					.fontWeight = Gfx::Font::Weight::Normal(),
+					.fontSize = Gfx::Length::Emphemeral(26.0 / 30.0)
+				},
+				Text
+				{
+					.color = Gfx::Color::Gray(0.65),
+					.textAlign = Anim::Interpolated<Text::TextAlign>(Text::TextAlign::left),
+					.backgroundColor = Gfx::Color(),
+					.numberFormat = ::Text::NumberFormat::prefixed,
+					.maxFractionDigits = 3,
+					.numberScale = ::Text::NumberScale{}
+				},
+			},
 			.tooltip = {
 				Font
 				{
@@ -466,7 +516,8 @@ Chart Chart::def()
 			.logo = {
 				Padding
 				{
-					.paddingTop = Gfx::Length::Relative(0.475),.paddingRight = Gfx::Length::Relative(0.4),
+					.paddingTop = Gfx::Length::Relative(0.475),
+					.paddingRight = Gfx::Length::Relative(0.4),
 					.paddingBottom = Gfx::Length::Relative(0.475),
 					.paddingLeft = Gfx::Length::Relative(0.4)
 				},