+++ tools/preset-typeschema-gen/preset-typeschema-gen.js
@@ -94,6 +94,14 @@ function genSchema(presets)
 						type: 'string',
 						nullable: true	
 					},
+					subtitle: {
+						type: 'string',
+						nullable: true
+					},
+					caption: {
+						type: 'string',
+						nullable: true
+					},
 					reverse: {
 						type: 'boolean'
 					},